+++ RELEASE-NOTES.md
@@ -11,6 +11,16 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ---
 
+## v2019.01.4
+
+- CHN: Add REGISTRANT_OPERATOR_ID and INQUIRER_OPERATOR_ID to the convicted report.
+- CHN: Remove CrimeEvent from the convicted report.
+- NEW: Add "Clear fields" button in Convicted-Report-Inquiry-By-Search-Criteria workflow.
+- ENH: Upgrade gradle to 5.1.1.
+- FIX: Disable wrapping top menu titles.
+
+---
+
 ## v2019.01.3
 
 - NEW: Add Delete-Convicted-Report workflow.